+++ api/Onsharp/Entities/Factory/Text3DFactory.cs
@@ -5,7 +5,7 @@
     /// </summary>
     internal class Text3DFactory : IEntityFactory<Text3D>
     {
-        public Text3D Create(long id)
+        public Text3D Create(int id)
         {
             return new Text3D(id);
         }
